Output;
Enter a positive number: 12321
The reverse of the number is: 12321
The number is a palindrome.

Enter a positive number: 12331
The reverse of the number is: 13321
The number is not a palindrome.
Sample Input:
Case = 1
String = MADAM Sample Output: Palindrome
Test cases:
a.​MONEY:
    Palindrome
b.​5678765:
      Palindrome
c. MALAY12321ALAM:
      Not a palindrome
d. MALAYALAM:
       Palindrome
e.​1234.4321:
      Not a Palindrome